Update PS3 Gentoo post, remove Vim cheatsheet, update URLs
Replaced the content of the PS3 Gentoo post with a link to a PDF file, deleted the Vim cheatsheet post, and updated all site URLs from localhost to stuffbymax.me in HTML and XML files. Also updated the last update date in hugo.toml and added a new image for the PS3 Gentoo tutorial.